code feedback updated
