--- src/components/Header/HeaderContainer.js
@@ -1,8 +1,0 @@
-import HeaderPresenter from "./HeaderPresenter";
-
-
-const HeaderContainer = () => {
-  return (<HeaderPresenter />);
-};
-
-export default HeaderContainer;
