--- src/components/Header/index.js
@@ -1,3 +1,0 @@
-import HeaderContainer from "./HeaderContainer";
-
-export default HeaderContainer;
