+++ src/components/MainApp/HeaderContainer.js
@@ -1,0 +1,8 @@
+import HeaderPresenter from "./HeaderPresenter";
+
+
+const HeaderContainer = () => {
+  return (<HeaderPresenter />);
+};
+
+export default HeaderContainer;
