+++ src/components/MainApp/index.js
@@ -1,0 +1,3 @@
+import HeaderContainer from "./HeaderContainer";
+
+export default HeaderContainer;
